+++ README.md
@@ -4,3 +4,17 @@ Uses Node.js
 
 Run "npm install" to get all the dependencies
 Run "npm start" to open up a webpage with the project at localhost:8080
+
+IDEAS:
+~local multiplayer game
+~one player controls with wasd, one person with arrows
+~attack with some sort of attack button
+~character select screen, with different stats (strength, speed, cool stuff)
+~have it in some kind of arena
+~jousting?
+    -Maybe have something where each player mashes a button to speed up the horse by a certain value
+    -Once they get close enough perhaps have them time a button press to lower their lance
+    -Whoever has better timing does more damage to the other person
+~possible upgrades
+~swordfighting on horses?
+~maybe implement some kind of story at the end
