+++ src/config.js
@@ -1,16 +1,18 @@
-export default = {
-  type: Phaser.AUTO,
-  width: 800,
-  height: 600,
-  physics: {
-  	default: "arcade",
-  	arcade: {
-  		gravity: {y: 0},
-  		debug: true
-  	}
-  },
-  scene: {
-    preload: preload,
-    create: create
-  }
+import TitleScreen from "./title.js"
+import GameScreen from "./game.js"
+
+const config = {
+    type: Phaser.AUTO,
+    width: 800,
+    height: 600,
+    scene: [TitleScreen, GameScreen],
+    physics: {
+    	 default: "arcade",
+    	 arcade: {
+    		gravity: {y: 0},
+    		debug: true
+    	},
+    }
 };
+
+export {config};
